Move Sample to frequenz.sdk.timeseries
Now that we have a timeseries package it makes more sense to have the
Sample class living there.
