+++ src/frequenz/sdk/actor/data_sourcing/microgrid_api_source.py
@@ -14,7 +14,7 @@ from typing import Any, Callable, Dict, List, Optional, Tuple
 from frequenz.channels import Receiver, Sender
 
 from frequenz.sdk.actor import ChannelRegistry
-from frequenz.sdk.data_pipeline import ComponentMetricId, ComponentMetricRequest, Sample
+from frequenz.sdk.data_pipeline import ComponentMetricId, ComponentMetricRequest
 from frequenz.sdk.microgrid import (
     BatteryData,
     ComponentCategory,
@@ -23,6 +23,7 @@ from frequenz.sdk.microgrid import (
     MeterData,
     microgrid_api,
 )
+from frequenz.sdk.timeseries import Sample
 
 _MeterDataMethods: Dict[ComponentMetricId, Callable[[MeterData], float]] = {
     ComponentMetricId.ACTIVE_POWER: lambda msg: msg.active_power,
