+++ src/frequenz/sdk/timeseries/sample.py
@@ -1,0 +1,25 @@
+"""Common types for the Data Pipeline.
+
+Copyright
+Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+
+License
+MIT
+"""
+
+from dataclasses import dataclass
+from datetime import datetime
+from typing import Optional
+
+
+@dataclass(frozen=True)
+class Sample:
+    """A measurement taken at a particular point in time.
+
+    The `value` could be `None` if a component is malfunctioning or data is
+    lacking for another reason, but a sample still needs to be sent to have a
+    coherent view on a group of component metrics for a particular timestamp.
+    """
+
+    timestamp: datetime
+    value: Optional[float] = None
